test framework

+++ std/sys/types.h
@@ -11,7 +11,7 @@ using i16 = int16_t;
 using u16 = uint16_t;
 
 using i32 = int32_t;
-using u32 = uint64_t;
+using u32 = uint32_t;
 
 using i64 = int64_t;
 using u64 = uint64_t;
